Ver 2.0.1.06 Query fixes for Version 2 Files modified - [PartPSRPipe,Job1111,webapp,.wmproject.properties]

+++ services/Job1111/src/com/psabackoffice/job1111/models/query/GetBidWorkHistoryEquipResponse.java
@@ -47,7 +47,7 @@ public class GetBidWorkHistoryEquipResponse implements Serializable {
     private String notes;
 
     @ColumnAlias("Rework")
-    private Integer rework;
+    private Short rework;
 
     public Integer getSubmissionId() {
         return this.submissionId;
@@ -129,11 +129,11 @@ public class GetBidWorkHistoryEquipResponse implements Serializable {
         this.notes = notes;
     }
 
-    public Integer getRework() {
+    public Short getRework() {
         return this.rework;
     }
 
-    public void setRework(Integer rework) {
+    public void setRework(Short rework) {
         this.rework = rework;
     }
